requirements.txt
livekit-agents>=0.8.0
livekit-api>=0.6.0
livekit-plugins-google>=0.6.0
livekit-plugins-silero>=0.6.0
livekit-plugins-noise-cancellation
python-dotenv>=1.0.0
